Refactoring `intersection.c` to utilize native vector operations.
This increases precision and performance, but requires extensive
refactoring. This aims to fix
https://github.com/fiedl/hole-ice-study/issues/28.

On the plus side, this helps removing the un-needed old code mentioned
in
https://github.com/fiedl/hole-ice-study/issues/2#issuecomment-362939892.

Using native vector operations also requires to re-check the
xy-projection. Fortunately, there are still tests in place for that.
See: https://github.com/fiedl/hole-ice-study/issues/25.